ed-button: export a single component per file
